+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Remove extractor's `id` field, use `identifier` field instead
+
+### Removed
+- Remove extractor's `id` field
 
 ## [4.2.0] - 2023-02-13
 ### Added